+++ hand_coded_sources/C_VDD_hashmap.cpp
@@ -63,7 +63,7 @@ void C_VDD_hashmap::MyBlockavltree_element_for_collision::operator delete (void 
   smAllocInfo.mFreeList = p ;
   smAllocInfo.mAllocatedObjectCount -- ;
   if (smAllocInfo.mAllocatedObjectCount == 0) {
-    for (int32_t i=0 ; i<smAllocInfo.mAllocatedBlockCount ; i++) {
+    for (size_t i=0 ; i<smAllocInfo.mAllocatedBlockCount ; i++) {
       delete [] smAllocInfo.mAllocatedBlockList [i] ;
     }
     delete [] smAllocInfo.mAllocatedBlockList ;
@@ -83,9 +83,9 @@ void C_VDD_hashmap::MyBlockavltree_element_for_collision::operator delete (void 
 void C_VDD_hashmap::C_TreeForCollision::allocBlock (void) {
 //--- Realloc block list ?
   if (smAllocInfo.mAllocatedBlockListSize <= smAllocInfo.mAllocatedBlockCount) {
-    const int32_t newSize = smAllocInfo.mAllocatedBlockCount + 1024 ;
+    const size_t newSize = smAllocInfo.mAllocatedBlockCount + 1024 ;
     char ** newBlockList = new char * [newSize] ;
-    for (int32_t i=0 ; i<smAllocInfo.mAllocatedBlockCount ; i++) {
+    for (size_t i=0 ; i<smAllocInfo.mAllocatedBlockCount ; i++) {
       newBlockList [i] = smAllocInfo.mAllocatedBlockList [i] ;
     }
     delete [] smAllocInfo.mAllocatedBlockList ;
@@ -96,15 +96,15 @@ void C_VDD_hashmap::C_TreeForCollision::allocBlock (void) {
   smAllocInfo.mAllocatedBlockList [smAllocInfo.mAllocatedBlockCount] = new char [BLOCK_SIZE] ;
   char * ptr = & (smAllocInfo.mAllocatedBlockList [smAllocInfo.mAllocatedBlockCount] [0]) ;
   smAllocInfo.mAllocatedBlockCount ++ ;
-  int32_t blockSize = BLOCK_SIZE ;
+  size_t blockSize = BLOCK_SIZE ;
   const int32_t ALIGNMENT = 32 ;
 //--- Align pointer
-  if ((((intptr_t) ptr) % ALIGNMENT) != 0) {
-    ptr = (char *) (((((intptr_t) ptr) / ALIGNMENT) + 1) * ALIGNMENT) ;
+  if ((((size_t) ptr) % ALIGNMENT) != 0) {
+    ptr = (char *) (((((size_t) ptr) / ALIGNMENT) + 1) * ALIGNMENT) ;
     blockSize -= ALIGNMENT ;
   }
-  const int32_t nbNewObjects = blockSize / ((int32_t) sizeof (MyBlockavltree_element_for_collision)) ;
-  for (int32_t i=0 ; i<nbNewObjects ; i++) {
+  const size_t nbNewObjects = blockSize / sizeof (MyBlockavltree_element_for_collision) ;
+  for (size_t i=0 ; i<nbNewObjects ; i++) {
     MyBlockavltree_element_for_collision * newObjectPtr = (MyBlockavltree_element_for_collision *) ptr ;
     newObjectPtr->mPtrToSup = smAllocInfo.mFreeList ;
     smAllocInfo.mFreeList = newObjectPtr ;
@@ -139,7 +139,7 @@ C_VDD_hashmap::C_TreeForCollision::~C_TreeForCollision (void) {
 //                                                                           *
 //---------------------------------------------------------------------------*
 
-uint32_t C_VDD_hashmap::getAllocatedSizeInBytes (void) {
+size_t C_VDD_hashmap::getAllocatedSizeInBytes (void) {
   return (uint32_t) (BLOCK_SIZE * smAllocInfo.mAllocatedBlockCount) ;
 }
 
@@ -462,9 +462,9 @@ unmarkAllObjects (void) {
 
 //---------------------------------------------------------------------------*
 
-int32_t C_VDD_hashmap::C_TreeForCollision::
+size_t C_VDD_hashmap::C_TreeForCollision::
 internalMarkedNodeCount (const MyBlockavltree_element_for_collision * const inElement) const {
-  int32_t result = 0 ;
+  size_t result = 0 ;
   if (inElement != NULL) {
     if (inElement->mInfo.isMarked ()) {
       result ++ ;
@@ -477,7 +477,7 @@ internalMarkedNodeCount (const MyBlockavltree_element_for_collision * const inEl
 
 //---------------------------------------------------------------------------*
 
-int32_t C_VDD_hashmap::C_TreeForCollision::
+size_t C_VDD_hashmap::C_TreeForCollision::
 getMarkedNodesCount (void) const {
   return internalMarkedNodeCount (mRoot) ;
 }
@@ -521,7 +521,7 @@ search (const C_VDD::cVDDnodeInfo & inInfo) {
 
 uint32_t C_VDD_hashmap::sweepUnmarkedObjects (void) {
   uint32_t sweepedNodes = 0 ;
-  for (int32_t i=0 ; i<mEntryCurrentCount ; i++) {
+  for (size_t i=0 ; i<mEntryCurrentCount ; i++) {
     sweepedNodes += mMapArray [i].sweepUnmarkedObjects () ;
   }
   return sweepedNodes ;
@@ -537,10 +537,10 @@ uint32_t C_VDD_hashmap::getMapSizeInBytes (void) const {
 
 void C_VDD_hashmap
 ::reallocMap (const int32_t inNewSize) {
-  const int32_t newSize = (inNewSize < 1) ? 1 : (int32_t) getPrimeGreaterThan ((uint32_t) inNewSize) ;
+  const size_t newSize = (inNewSize < 1) ? 1 : (size_t) getPrimeGreaterThan ((uint32_t) inNewSize) ;
   if (newSize != mEntryCurrentCount) {
     C_TreeForCollision * newMapArray = (newSize > 1) ? new C_TreeForCollision [newSize] : (& mMinimumMapArray) ;
-    for (int32_t i=0 ; i<mEntryCurrentCount ; i++) {
+    for (size_t i=0 ; i<mEntryCurrentCount ; i++) {
       mMapArray [i].transfertElementsInNewMapArray (newMapArray, (uint32_t) newSize) ;
     }
     if (mEntryCurrentCount > 1) {
@@ -554,16 +554,16 @@ void C_VDD_hashmap
 //---------------------------------------------------------------------------*
 
 void C_VDD_hashmap::unmarkAllObjects (void) {
-  for (int32_t i=0 ; i<mEntryCurrentCount ; i++) {
+  for (size_t i=0 ; i<mEntryCurrentCount ; i++) {
     mMapArray [i].unmarkAllObjects () ;
   }
 }
 
 //---------------------------------------------------------------------------*
 
-int32_t C_VDD_hashmap::getMarkedNodesCount (void) const {
-  int32_t count = 0 ;
-  for (int32_t i=0 ; i<mEntryCurrentCount ; i++) {
+size_t C_VDD_hashmap::getMarkedNodesCount (void) const {
+  size_t count = 0 ;
+  for (size_t i=0 ; i<mEntryCurrentCount ; i++) {
     count += mMapArray [i].getMarkedNodesCount () ;
   }
   return count ;
