+++ hand_coded_sources/C_VDD_hashmap.h
@@ -62,7 +62,7 @@ class C_VDD_hashmap {
   public : uint32_t getMapSizeInBytes (void) const ;
 
 //--- Get marked nodes count
-  public : int32_t getMarkedNodesCount (void) const ;
+  public : size_t getMarkedNodesCount (void) const ;
 
 //--- Sweep unmarked objects
   public : uint32_t sweepUnmarkedObjects (void) ;
@@ -71,14 +71,14 @@ class C_VDD_hashmap {
   public : void unmarkAllObjects (void) ;
 
 //--- Current Entry count
-  protected : int32_t mEntryCurrentCount ;
-  public : inline int32_t getHashMapEntryCount (void) const { return mEntryCurrentCount ; }
+  protected : size_t mEntryCurrentCount ;
+  public : inline size_t getHashMapEntryCount (void) const { return mEntryCurrentCount ; }
 
 //--- Get node size (in bytes)
-  public : static uint32_t getNodeSize (void) { return (uint32_t) sizeof (MyBlockavltree_element_for_collision) ; }
+  public : static size_t getNodeSize (void) { return sizeof (MyBlockavltree_element_for_collision) ; }
 
 //--- Get allocated size (in bytes)
-  public : static uint32_t getAllocatedSizeInBytes (void) ;
+  public : static size_t getAllocatedSizeInBytes (void) ;
 
 //--- No copy
   private : C_VDD_hashmap (const C_VDD_hashmap &) ;
@@ -100,7 +100,7 @@ class C_VDD_hashmap {
   public : inline C_VDD::cVDDnodeInfo * search (const C_VDD::cVDDnodeInfo & inInfo) ;
 
 //--- Get marked nodes count
-  public : int32_t getMarkedNodesCount (void) const ;
+  public : size_t getMarkedNodesCount (void) const ;
 
 //--- Sweep unmarked objects
   public : uint32_t sweepUnmarkedObjects (void) ;
@@ -140,7 +140,7 @@ class C_VDD_hashmap {
                                            bool & outExtension,
                                            bool & outInsertionPerformed) ;
 
-  protected : int32_t internalMarkedNodeCount (const MyBlockavltree_element_for_collision * const inElement) const ;
+  protected : size_t internalMarkedNodeCount (const MyBlockavltree_element_for_collision * const inElement) const ;
 
   protected : uint32_t internalRecursiveSweep (MyBlockavltree_element_for_collision * inElement) ;
 
@@ -183,11 +183,11 @@ class C_VDD_hashmap {
 //------------------- MyBlockavltree_element_for_collision embedded class
   private : class cAllocInfo {
     public : char * * mAllocatedBlockList ;
-    public : int32_t mAllocatedBlockListSize ;
-    public : int32_t mAllocatedBlockCount ;
+    public : size_t mAllocatedBlockListSize ;
+    public : size_t mAllocatedBlockCount ;
     public : MyBlockavltree_element_for_collision * mFreeList ;
-    public : int32_t mAllocatedObjectCount ;
-    public : int32_t mCreatedObjectCount ;
+    public : size_t mAllocatedObjectCount ;
+    public : size_t mCreatedObjectCount ;
     public : cAllocInfo (void) :
     mAllocatedBlockList (NULL),
     mAllocatedBlockListSize (0),
@@ -205,10 +205,10 @@ class C_VDD_hashmap {
   protected : C_TreeForCollision * mMapArray ;
 
 //--- Get created element count
-  public : static int32_t getCreatedObjectCount (void) { return smAllocInfo.mCreatedObjectCount ; }
+  public : static size_t getCreatedObjectCount (void) { return smAllocInfo.mCreatedObjectCount ; }
 
 //--- Get currently used element count
-  public : static int32_t getCurrentObjectCount (void) { return smAllocInfo.mAllocatedObjectCount ; }
+  public : static size_t getCurrentObjectCount (void) { return smAllocInfo.mAllocatedObjectCount ; }
 
 //--- Allocation info (static variable)
   protected : static cAllocInfo smAllocInfo ;
